Functional tests (company packages, shipping)

+++ functional-tests/tests/basic_tests.rs
@@ -2243,12 +2243,12 @@ fn create_update_delete_warehouse() {
         },
         location: Some(update_warehouse::GeoPointInput {
             x: 42.0,
-            y: 666.0,
+            y: 90.0,
             ..update_warehouse::default_geo_point_input()
         }),
         name: Some("New name".to_string()),
         ..update_warehouse::default_update_warehouse_input()
-    }).expect("Cannot get data from update_warehouse");
+    }).expect("Cannot get data from update_warehouse").expect("Empty data from update_warehouse");
 
     assert_eq!(updated_warehouse.name, Some("New name".to_string()));
     assert_eq!(updated_warehouse.address_full.country, Some("Russian Federation".to_string()));
@@ -2256,18 +2256,39 @@ fn create_update_delete_warehouse() {
     assert_eq!(updated_warehouse.address_full.administrative_area_level2, Some("Moscow".to_string()));
     let location = updated_warehouse.location.expect("Cannot get location data from update_warehouse");
     assert_eq!(location.x, 42.0);
-    assert_eq!(location.y, 666.0);
+    assert_eq!(location.y, 90.0);
 
     let deleted_warehouse_id = delete_warehouse(&mut context, token.clone(), id.clone())
         .expect("Cannot get data from delete_warehouse")
-        .expect("Cannot get data from delete_warehouse");
+        .expect("Empty data from delete_warehouse");
+
+    assert_eq!(deleted_warehouse_id, id);
+
+    // negative cases
+    let update_deleted_warehouse = update_warehouse(&mut context, token.clone(), update_warehouse::UpdateWarehouseInput {
+        id: id.clone(),
+        address_full: update_warehouse::AddressInput {
+            country: Some("Russian Federation".to_string()),
+            administrative_area_level1: Some("Moscow Region".to_string()),
+            administrative_area_level2: Some("Moscow".to_string()),
+            ..update_warehouse::default_address_input()
+        },
+        location: Some(update_warehouse::GeoPointInput {
+            x: 42.0,
+            y: 90.0,
+            ..update_warehouse::default_geo_point_input()
+        }),
+        name: Some("New name".to_string()),
+        ..update_warehouse::default_update_warehouse_input()
+    });
+    if update_deleted_warehouse.is_ok() && update_deleted_warehouse.unwrap().is_some() {
+        panic!("Should not be able to update deleted warehouse");
+    }
 
     let deleted_twice = delete_warehouse(&mut context, token.clone(), id.clone());
     if deleted_twice.is_ok() && deleted_twice.unwrap() != None {
         panic!("Should not be able to delete the same warehouse twice");
     }
-
-    assert_eq!(deleted_warehouse_id, id);
 }
 
 #[test]
@@ -2279,19 +2300,23 @@ fn create_update_delete_package() {
             "RUS".to_string(),
             "USA".to_string()
         ],
-        ..create_package::default_new_packages_input()
+        ..create_package::default_graphql_request_input()
     }).expect("Cannot get data from create_package");
 
-    assert_eq!(new_package.name, "Initial name".to_string());
-    assert_eq!(new_package.max_size, 1000);
-    assert_eq!(new_package.min_size, 100);
-    assert_eq!(new_package.max_weight, 3000);
-    assert_eq!(new_package.min_weight, 300);
+    let get_package = context.request(get_package::GetPackageInput { id: new_package.raw_id })
+        .expect("Cannot get data from get_package")
+        .expect("Empty data from get_package");
+
+    assert_eq!(get_package.name, "Initial name".to_string());
+    assert_eq!(get_package.max_size, 1000);
+    assert_eq!(get_package.min_size, 100);
+    assert_eq!(get_package.max_weight, 3000);
+    assert_eq!(get_package.min_weight, 300);
 
     // deliveries
-    assert_eq!(new_package.deliveries_to.len(), 1);
+    assert_eq!(get_package.deliveries_to.len(), 1);
 
-    let xal = new_package.deliveries_to.first().expect("Cannot get delivery info");
+    let xal = get_package.deliveries_to.first().expect("Cannot get delivery info");
     assert_eq!(xal.level, 0);
     assert_eq!(xal.label, "All".to_string());
     assert_eq!(xal.alpha3, "XAL".to_string());
@@ -2324,19 +2349,144 @@ fn create_update_delete_package() {
         min_size: Some(101),
         max_weight: Some(3001),
         min_weight: Some(301),
-        ..update_package::default_update_packages_input()
+        ..update_package::default_graphql_request_input()
     }).expect("Cannot get data from update_package");
 
-    assert_eq!(updated_package.name, "New name".to_string());
-    assert_eq!(updated_package.max_size, 1001);
-    assert_eq!(updated_package.min_size, 101);
-    assert_eq!(updated_package.max_weight, 3001);
-    assert_eq!(updated_package.min_weight, 301);
+    let get_package = context.request(get_package::GetPackageInput { id: new_package.raw_id })
+        .expect("Cannot get data from get_package")
+        .expect("Empty data from get_package");
+
+    assert_eq!(get_package.name, "New name".to_string());
+    assert_eq!(get_package.max_size, 1001);
+    assert_eq!(get_package.min_size, 101);
+    assert_eq!(get_package.max_weight, 3001);
+    assert_eq!(get_package.min_weight, 301);
 
     delete_package(&mut context, new_package.raw_id).expect("Cannot get deleted package");
+
+    // negative cases
+    let get_package = context.request(get_package::GetPackageInput { id: new_package.raw_id });
+    if get_package.is_ok() && get_package.unwrap().is_some() {
+        panic!("Should not be able to get deleted package");
+    }
+
     if delete_package(&mut context, new_package.raw_id).is_ok() {
         panic!("Should not be able to delete the same package twice");
     }
+}
+
+#[test]
+fn create_delete_company_package() {
+    let mut context = TestContext::new();
+    context.as_superadmin();
+
+    let new_package = create_package(&mut context, create_package::NewPackagesInput {
+        name: "Initial name".to_string(),
+        deliveries_to: vec![
+            "RUS".to_string(),
+            "USA".to_string()
+        ],
+        ..create_package::default_graphql_request_input()
+    }).expect("Cannot get data from create_package");
+
+    let new_company = context
+        .request(create_delivery_company::NewCompanyInput {
+            name: "Test company".to_string(),
+            label: "TEST".to_string(),
+            description: Some("Test description".to_string()),
+            deliveries_from: vec!["RUS".to_string()],
+            logo: "test loge URL".to_string(),
+            ..create_delivery_company::default_create_company_input()
+        })
+        .expect("Cannot get data from create_delivery_company");
+
+    let company_package = add_package_to_company(&mut context, add_package_to_company::NewCompaniesPackagesInput {
+        company_id: new_company.raw_id,
+        package_id: new_package.raw_id,
+        ..add_package_to_company::default_graphql_request_input()
+    }).expect("Cannot get data from add_package_to_company");
+
+    let company = company_package.company.expect("Cannot get company data from add_package_to_company");
+    let package = company_package.package.expect("Empty data from add_package_to_company");
+
+    assert_eq!(company.label, new_company.label);
+    assert_eq!(company.name, new_company.name);
+    assert_eq!(package.name, new_package.name);
+
+    let company_package = context.request(get_company_package::GetCompanyPackageInput { id: company_package.raw_id })
+        .expect("Cannot get data from get_company_package")
+        .expect("Empty data from get_company_package");
+    assert_eq!(company_package.company_id, new_company.raw_id);
+    assert_eq!(company_package.package_id, new_package.raw_id);
+
+    delete_company_package(&mut context, new_company.raw_id, new_package.raw_id)
+        .expect("Cannot get data from delete_company_package");
+
+    // negative cases
+    let deleted_company_package = context.request(get_company_package::GetCompanyPackageInput { id: company_package.raw_id });
+    if deleted_company_package.is_ok() && deleted_company_package.unwrap().is_some() {
+        panic!("Should not be able to get deleted company package");
+    }
+
+    if delete_company_package(
+        &mut context,
+        new_company.raw_id,
+        new_package.raw_id
+    ).is_ok() {
+        panic!("Should not be able to delete the same company package twice");
+    }
+}
+
+#[test]
+fn upsert_shipping() {
+    let mut context = TestContext::new();
+
+    let (user, token, store, category, base_product) = set_up_base_product(&mut context).expect("Cannot get data from set_up_base_product");
+
+    context.set_bearer(token);
+    let warehouse_payload = create_warehouse::CreateWarehouseInput {
+        name: Some("Warehouse".to_string()),
+        store_id: store.raw_id,
+        address_full: create_warehouse::AddressInput {
+            country: Some("Russian Federation".to_string()),
+            country_code: Some("RUS".to_string()),
+            ..create_warehouse::default_address_input()
+        },
+        ..create_warehouse::default_graphql_request_input()
+    };
+    let warehouse = context.request(warehouse_payload).expect("Cannot get data from create_warehouse");
+
+    context.as_superadmin();
+    let upsert_shipping_payload = upsert_shipping::NewShippingInput {
+        store_id: store.raw_id,
+        base_product_id: base_product.raw_id,
+        ..upsert_shipping::default_graphql_request_input()
+    };
+    let upsert_shipping = context.request(upsert_shipping_payload).expect("Cannot get data from upsert_shipping");
+}
+
+fn add_package_to_company(
+    context: &mut TestContext,
+    payload: add_package_to_company::NewCompaniesPackagesInput
+) -> Result<add_package_to_company::GraphqlRequestOutput, FailureError> {
+    context.as_superadmin();
+    let company_package = context.request(payload)?;
+    context.clear_bearer();
+    Ok(company_package)
+}
+
+fn delete_company_package(
+    context: &mut TestContext,
+    company_id: i64,
+    package_id: i64
+) -> Result<delete_company_package::GraphqlRequestOutput, FailureError> {
+    context.as_superadmin();
+    let company_package = context.request(delete_company_package::DeleteCompanyPackageInput {
+        company_id,
+        package_id
+    })?;
+    context.clear_bearer();
+    Ok(company_package)
 }
 
 fn create_package(
@@ -2380,7 +2530,7 @@ fn set_up_warehouse(
     let warehouse_payload = create_warehouse::CreateWarehouseInput {
         name: Some("Initial name".to_string()),
         store_id: store.raw_id,
-        ..create_warehouse::default_create_warehouse_input()
+        ..create_warehouse::default_graphql_request_input()
     };
     let warehouse = context.request(warehouse_payload)?;
     context.clear_bearer();
@@ -2392,13 +2542,17 @@ fn update_warehouse(
     token: String,
     input: update_warehouse::UpdateWarehouseInput
 ) -> Result<(
-    get_store::RustGetStoreStoreWarehouses
+    Option<get_store::RustGetStoreStoreWarehouses>
 ), FailureError> {
     context.set_bearer(token);
-    let updated_warehouse = context.request(input)?.expect("Cannot get data from update_warehouse");
+    let updated_warehouse = context.request(input)?;
+    if updated_warehouse.is_none() {
+        return Ok(None);
+    }
+    let updated_warehouse = updated_warehouse.unwrap();
     let store = context.get_store(updated_warehouse.store_id)?.store.expect("Cannot get data from get_store");
     context.clear_bearer();
-    Ok(store.warehouses.into_iter().find(|x| x.id == updated_warehouse.id).expect("Cannot get warehouse data from update_warehouse"))
+    Ok(store.warehouses.into_iter().find(|x| x.id == updated_warehouse.id))
 }
 
 fn delete_warehouse(context: &mut TestContext, token: String, id: String) -> Result<Option<String>, FailureError> {
